+++ packages/web-component/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @mindfiredigital/page-builder-web-component
 
+## 5.0.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mindfiredigital/page-builder@1.20.0
+
+## 5.0.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mindfiredigital/page-builder@1.19.1
+
 ## 5.0.12
 
 ### Patch Changes